# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WedWithMe REST API",
    "version": "1.0.0",
    "description": "Autogenerated OpenAPI specification for the WedWithMe Next.js Backend."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API Environment"
    }
  ],
  "paths": {
...
    "/messages": {
      "get": {
        "summary": "GET /messages",
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "POST /messages",
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/payments/webhook": {
      "post": {
        "summary": "POST /payments/webhook",
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/vendors": {
      "get": {
        "summary": "GET /vendors",
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}